pyproject.toml
[project]
name = "firefly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.6",
    "djangorestframework>=3.15.2",
    "django-cors-headers>=4.3.1,<5.0.0",
    "djangorestframework-jsonapi[openapi]>=6.1.0,<7.0.0",
    "djangorestframework-simplejwt<6.0.0,>=5.3.1",
    "openai>=1.64.0",
    "psycopg2-binary>=2.9.10",
    "pandas>=2.2.3",
    "dotenv>=0.9.9",
    "celery[redis]>=5.4.0",
    "celery-singleton>=0.3.1",
    "python-dotenv>=1.0.1",
    "django-celery-beat>=2.7.0",
    "celery[redis]>=5.4.0",
    "celery-singleton>=0.3.1",
    "python-dotenv>=1.0.1",
    "django-celery-beat>=2.7.0",
]
